Refactor extract archive to account for folder inside zip

+++ utils.py
@@ -84,8 +84,12 @@ Xposed -> {xposed_url}@{previous_version(2, release)}
     send_message(message, tg_token, tg_chat_id, tg_thread_id)
 
 
-def extract_archive(zip_path: str, dir_path: str, file_path: str, regex: str):
-    shutil.unpack_archive(zip_path, dir_path)
+def extract_archive(zip_path: str, dir_path: str, file_path: str, regex: str, keep_dir: bool, out_dir: str = None, folders: str = None):
+    if out_dir is None and folders is None:
+        shutil.unpack_archive(zip_path, dir_path)
+    else:
+        shutil.unpack_archive(zip_path, out_dir)
+        os.rename(f"{out_dir}.lstrip('/')/{folders}", dir_path)
 
     os.remove(zip_path)
 
@@ -95,7 +99,8 @@ def extract_archive(zip_path: str, dir_path: str, file_path: str, regex: str):
                 os.unlink(file_path)
             os.rename(f"{dir_path}/{i}", file_path)
 
-    shutil.rmtree(dir_path)
+    if keep_dir is False:
+        shutil.rmtree(dir_path)
 
 
 def download(link, out, headers=None):
